# Speed up frame (un)masking with a C extensions.