--- a/Makefile
+++ b/Makefile
@@ -10,6 +10,6 @@ coverage:
 	python -m coverage html
 
 clean:
-	find . -name '*.pyc' -delete
+	find . -name '*.pyc' -o -name '*.so' -delete
 	find . -name __pycache__ -delete
 	rm -rf .coverage build compliance/reports dist docs/_build htmlcov MANIFEST README websockets.egg-info