--- a/tox.ini
+++ b/tox.ini
@@ -1,16 +1,39 @@
 [tox]
-envlist = py33,py34,py35,py36,coverage,flake8,isort
+envlist = {py33,py34,py35,py36}{,-speedups},coverage,flake8,isort
 
 [testenv]
+commands =
+    ; Unfortunately tox has no support for building C extensions.
+    ; Do it manually in the git checkout - that's where tests run.
+
+    ; Remove any existing compiled extension.
+    sh -c 'rm -f websockets/*.so'
+
+    ; Before testing with speedups, compile the extension.
+    speedups: python setup.py --quiet build_ext --inplace
+
+    python -m unittest
+
+    ; After testing with speedups, remove the extension.
+    speedups: sh -c 'rm websockets/*.so'
+
 deps =
     py33: asyncio
-commands = python -m unittest
+whitelist_externals =
+    sh
 
 [testenv:coverage]
 commands =
+    ; Handle speedups as above.
+    sh -c 'rm -f websockets/*.so'
+    python setup.py --quiet build_ext --inplace
+
     python -m coverage erase
     python -m coverage run --branch --source=websockets -m unittest
     python -m coverage report --show-missing --fail-under=100
+
+    speedups: sh -c 'rm websockets/*.so'
+
 deps = coverage
 
 [testenv:flake8]