Thanks Messense Lv (@messense) and Matthieu Darbois (@mayeut) for proposing previous iterations of this patch in #175 and #211.