@aaugustin 2017-07-22
This PR is the third iteration after #175 and #211. It contains all the bits that I'm ready to merge at this point.

Given that I don't have a plan for building wheels on all platforms, sadly, I will have either to publish a wheel without the speedups or only a sdist. Only a sdist sounds best for now.

Benchmarks (Python 3.6 / 2,9 GHz Intel Core i7 / macOS) :

```python
>>> def benchmark(data_size, number):
...     # Previous Python version
...     old = timeit.timeit(
...         stmt='bytes(b ^ mask[i % 4] for i, b in enumerate(data))',
...         setup=f'import os; data = os.urandom({data_size}); mask = os.urandom(4);',
...         number=number,
...     )
...     # New Python version
...     new = timeit.timeit(
...         stmt='bytes(b ^ m for b, m in zip(data, itertools.cycle(mask)))',
...         setup=f'import itertools; import os; data = os.urandom({data_size}); mask = os.urandom(4);',
...         number=number,
...     )
...     # New C version
...     c = timeit.timeit(
...         stmt='apply_mask(data, mask)',
...         setup=f'import os; data = os.urandom({data_size}); mask = os.urandom(4); from websockets.speedups import apply_mask',
...         number=number,
...     )
...     print(f"* {data_size} bytes *")
...     print(f"Old: {data_size * number / old / 2**20:.1f}MB/sec")
...     print(f"New: {data_size * number / new / 2**20:.1f}MB/sec - {(old - new) / old * 100:.2f}% faster vs. old")
...     print(f"  C: {data_size * number / c / 2**20:.1f}MB/sec - {(new - c) / new * 100:.2f}% faster vs. new")
...
>>> benchmark(10, 1000000)
* 10 bytes *
Old: 4.0MB/sec
New: 4.4MB/sec - 8.42% faster vs. old
  C: 66.9MB/sec - 93.40% faster vs. new
>>> benchmark(1000, 10000)
* 1000 bytes *
Old: 5.7MB/sec
New: 8.6MB/sec - 32.97% faster vs. old
  C: 791.6MB/sec - 98.92% faster vs. new
>>> benchmark(100000, 100)
* 100000 bytes *
Old: 5.8MB/sec
New: 9.0MB/sec - 34.94% faster vs. old
  C: 939.7MB/sec - 99.04% faster vs. new
```