@messense 2017-07-23
For wheel building [cibuildwheel](https://github.com/joerick/cibuildwheel) might be useful if you are willing to do pypi release automatically in CI.