@aaugustin 2017-07-23
Oh, this project looks great! I filed #220 for this.