@aaugustin 2017-07-23
Benchmarks for AVX optimization (Python 3.6 / 2,9 GHz Intel Core i7 / macOS) :

Compiled with `CFLAGS=-march=native`: significant gain:

```
(websockets) myk@mYk:dev/websockets ‹c-speedup*›$ make clean
find . -name '*.pyc' -o -name '*.so' -delete
find . -name __pycache__ -delete
rm -rf .coverage build compliance/reports dist docs/_build htmlcov MANIFEST README websockets.egg-info
(websockets) myk@mYk:dev/websockets ‹c-speedup*›$ CFLAGS='-march=native' python setup.py build_ext --inplace
running build_ext
building 'websockets.speedups' extension
creating build
creating build/temp.macosx-10.12-x86_64-3.6
creating build/temp.macosx-10.12-x86_64-3.6/websockets
clang -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -march=native -I/Users/myk/.virtualenvs/websockets/include -I/Users/myk/.pyenv/versions/3.6.0/include/python3.6m -c websockets/speedups.c -o build/temp.macosx-10.12-x86_64-3.6/websockets/speedups.o
creating build/lib.macosx-10.12-x86_64-3.6
creating build/lib.macosx-10.12-x86_64-3.6/websockets
clang -bundle -undefined dynamic_lookup -L/usr/local/opt/readline/lib -L/usr/local/opt/readline/lib -L/usr/local/opt/openssl/lib -L/Users/myk/.pyenv/versions/3.6.0/lib -L/usr/local/opt/readline/lib -L/usr/local/opt/readline/lib -L/usr/local/opt/openssl/lib -L/Users/myk/.pyenv/versions/3.6.0/lib -march=native build/temp.macosx-10.12-x86_64-3.6/websockets/speedups.o -o build/lib.macosx-10.12-x86_64-3.6/websockets/speedups.cpython-36m-darwin.so
copying build/lib.macosx-10.12-x86_64-3.6/websockets/speedups.cpython-36m-darwin.so -> websockets
(websockets) myk@mYk:dev/websockets ‹c-speedup*›$ python
Python 3.6.0 (default, Jul  9 2017, 18:53:19)
[GCC 4.2.1 Compatible Apple LLVM 8.1.0 (clang-802.0.42)] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> from benchmark import benchmark
>>> benchmark(10, 100000)
* 10 bytes *
Old: 4.0MB/sec
New: 4.4MB/sec - 10.38% faster vs. old
  C: 66.1MB/sec - 93.28% faster vs. new
AVX: 64.8MB/sec - -2.12% faster vs. C
>>> benchmark(1000, 1000)
* 1000 bytes *
Old: 5.3MB/sec
New: 8.9MB/sec - 40.99% faster vs. old
  C: 812.9MB/sec - 98.90% faster vs. new
AVX: 3870.2MB/sec - 79.00% faster vs. C
>>> benchmark(100000, 10)
* 100000 bytes *
Old: 5.6MB/sec
New: 8.5MB/sec - 33.84% faster vs. old
  C: 848.2MB/sec - 99.00% faster vs. new
AVX: 30003.9MB/sec - 97.17% faster vs. C
>>>
```

Compiled without `CFLAGS`: no gain, the AVX branch isn't compiled:

```
(websockets) myk@mYk:dev/websockets ‹c-speedup*›$ make clean
find . -name '*.pyc' -o -name '*.so' -delete
find . -name __pycache__ -delete
rm -rf .coverage build compliance/reports dist docs/_build htmlcov MANIFEST README websockets.egg-info
(websockets) myk@mYk:dev/websockets ‹c-speedup*›$ python setup.py build_ext --inplace
running build_ext
building 'websockets.speedups' extension
creating build
creating build/temp.macosx-10.12-x86_64-3.6
creating build/temp.macosx-10.12-x86_64-3.6/websockets
clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -I/Users/myk/.virtualenvs/websockets/include -I/Users/myk/.pyenv/versions/3.6.0/include/python3.6m -c websockets/speedups.c -o build/temp.macosx-10.12-x86_64-3.6/websockets/speedups.o
creating build/lib.macosx-10.12-x86_64-3.6
creating build/lib.macosx-10.12-x86_64-3.6/websockets
clang -bundle -undefined dynamic_lookup -L/usr/local/opt/readline/lib -L/usr/local/opt/readline/lib -L/usr/local/opt/openssl/lib -L/Users/myk/.pyenv/versions/3.6.0/lib -L/usr/local/opt/readline/lib -L/usr/local/opt/readline/lib -L/usr/local/opt/openssl/lib -L/Users/myk/.pyenv/versions/3.6.0/lib build/temp.macosx-10.12-x86_64-3.6/websockets/speedups.o -o build/lib.macosx-10.12-x86_64-3.6/websockets/speedups.cpython-36m-darwin.so
copying build/lib.macosx-10.12-x86_64-3.6/websockets/speedups.cpython-36m-darwin.so -> websockets
(websockets) myk@mYk:dev/websockets ‹c-speedup*›$ python
Python 3.6.0 (default, Jul  9 2017, 18:53:19)
[GCC 4.2.1 Compatible Apple LLVM 8.1.0 (clang-802.0.42)] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> from benchmark import benchmark
>>> benchmark(10, 100000)
* 10 bytes *
Old: 4.0MB/sec
New: 4.1MB/sec - 4.04% faster vs. old
  C: 67.7MB/sec - 93.89% faster vs. new
AVX: 58.3MB/sec - -16.10% faster vs. C
>>> benchmark(1000, 1000)
* 1000 bytes *
Old: 5.7MB/sec
New: 9.1MB/sec - 37.73% faster vs. old
  C: 815.6MB/sec - 98.88% faster vs. new
AVX: 814.6MB/sec - -0.12% faster vs. C
>>> benchmark(100000, 10)
* 100000 bytes *
Old: 5.1MB/sec
New: 8.9MB/sec - 42.72% faster vs. old
  C: 942.6MB/sec - 99.05% faster vs. new
AVX: 943.5MB/sec - 0.09% faster vs. C
>>>
```